Add new flush: flush events or populate packet
Following this change[1], a report was filed[2] on the lttng-dev list
that subsequent snapshots of buffers with no new events would no
longer have events after a number of snapshots equal to the number of
sub-buffers.

The changes were initially made in order to make snapshots on channels
with few changing events intersectable by Babeltrace. By forcibly
closing the packet, the end timestamp represents when the snapshot
was taken and gives the information to the reader that no events were
produced between the last event and the end timestamp of the closed
packet.

Without the above information, the intersection between a busy stream
and a stream with no recent events would make the intersection window
too small.

However, this broke another use case where the content of subsequent
snapshots wouldn't change unless there were new events.

Proposed solution
-----------------

A new ioctl has been added to perform a new flush which may populate a
standalone packet with stream specific information if no changes have
been detected - LTTNG_KERNEL_ABI_RING_BUFFER_FLUSH_OR_POPULATE_PACKET.

In order to reach the conclusion if changes have been made, the following
test is performed within the new flush operation:

1. The produced position of the stream is sampled
2. The time is sampled
3. An active flush is done
4. The produced position of the stream is sampled again
5. If there there is no change in the produced position of the stream,
it can be inferred that no events have been written into the
sub-buffer. In this case, the packet is populated with the
time sampled in step 2 to give an indication to snapshot readers that
no changes have been made between the last event and the timestamp in
the populated packet.

LTTNG_KERNEL_ABI_RING_BUFFER_FLUSH_OR_POPULATE_PACKET takes a pointer to
a struct lttng_kernel_abi_ring_buffer_packet_flush_or_populate_packet_args.
The packet field of the structure should be set to a valid pointer to a
region of allocated memory at least as large as the maximum subbuffer
size for the stream. All other fields of the structure are output fields
that may be set during the ioctl.

[1]: https://github.com/lttng/lttng-tools/commit/f22dd89135f6b1749cd75735f04e3cb309da12c1
[2]: https://lists.lttng.org/pipermail/lttng-dev/2016-December/026805.html

Change-Id: Ie929a8128d9a2bc6bb53e78363dba00e776aa163
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>