Revert "Fix: powerpc builds with linux v6.13"
This reverts commit 9e9226618052111d31eaffd2fbb80af8e890a234.

The above commit introduced a work-around to allow powerpc builds to
complete with Linux v6.13. The change involves setting the build
environment variable `MO` to `KERNELDIR`. `MO` sets the output
directory for the files produced by the external module build.

One of the files written into the `MO` directory is a new
Makefile. This is done by the Kbuild infrastructure and not
lttng-modules build itself. When the Makefile in the `MO` directory is
written, it ends up containing an inclusion to itself as `MO` ==
`KERNELDIR`. This causes a subsequent `make` or `make module_install`
to fail with the following error:

    /root/lttng/linux/ppc64el-v6.14.0-rc3/sources/Makefile:5: *** Too many open files.  Stop.

There is a proposed patch[1] for the upstream kernel to fix the
problem upstream.

[1]: https://lore.kernel.org/lkml/20250218-buildfix-extmod-powerpc-v2-1-1e78fcf12b56@efficios.com/

Change-Id: Id39143e22f5498edf27ffbfac6245d944bb2bb97
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>