Add internal, minimal metadata JSON generation API
This patch adds an internal metadata JSON generation API to support
CTF 2 metadata stream production.

The JSON generation system keeps a stack of states to know whether or
not to prepend commas and to perform basic precondition validation.

Because this system uses the lttng_metadata_print*() functions
internally, all its functions accept a session pointer and the JSON
state is part of the `lttng_kernel_session_private` structure.

The stack is currently a static array with a maximum depth of 128.
I believe this is enough for the LTTng-modules use cases, but it could
be changed to a dynamically allocated stack later.

All the "internally public" lttng_kernel_mj_gen_*() generation functions
accept an optional key which must not be `NULL` if you're currently
directly filling an object, and `NULL` otherwise.

Here's a usage example, omitting error checks for clarity:

    lttng_kernel_mj_gen_open_obj(session, NULL);
        lttng_kernel_mj_gen_str(session, "type", "preamble");
        lttng_kernel_mj_gen_uint(session, "version", 2);
        lttng_kernel_mj_gen_open_array(session, "uuid");
            lttng_kernel_mj_gen_uint(session, NULL, 229);
            lttng_kernel_mj_gen_uint(session, NULL, 62);
            lttng_kernel_mj_gen_uint(session, NULL, 10);
            lttng_kernel_mj_gen_uint(session, NULL, 184);
            lttng_kernel_mj_gen_uint(session, NULL, 80);
            lttng_kernel_mj_gen_uint(session, NULL, 161);
            lttng_kernel_mj_gen_uint(session, NULL, 79);
            lttng_kernel_mj_gen_uint(session, NULL, 10);,
            lttng_kernel_mj_gen_uint(session, NULL, 183);
            lttng_kernel_mj_gen_uint(session, NULL, 16);
            lttng_kernel_mj_gen_uint(session, NULL, 181);
            lttng_kernel_mj_gen_uint(session, NULL, 240);
            lttng_kernel_mj_gen_uint(session, NULL, 187);
            lttng_kernel_mj_gen_uint(session, NULL, 169);
            lttng_kernel_mj_gen_uint(session, NULL, 196);
            lttng_kernel_mj_gen_uint(session, NULL, 172);
        lttng_kernel_mj_gen_close_array(session);
    lttng_kernel_mj_gen_close_obj(session);

Result (adding whitespaces for clarity):

    {
      "type": "preamble",
      "version": 2,
      "uuid": [
        229, 62, 10, 184, 80, 161, 79, 10,
        183, 16, 181, 240, 187, 169, 196, 172
      ]
    }

Change-Id: I4395c4f9cb60b2bfbe7ed69b01cd581f4ecdf27f
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>