Move common CTF metadata serialization code to `src/lttng-events.c`
This patch extracts code which would be common to any CTF metadata
serialization, moving it to common functions in `src/lttng-events.c`.
The goal is to avoid future redundancy when writing the CTF 2 metadata
serialization code.

It also adds lttng_channel_metadata_statedump_ctf_1_8() which is
equivalent to what _lttng_channel_metadata_statedump() was. This is
because serializing the metadata of a channel is both a CTF 1.8 and a
CTF 2 (upcoming) operation, as you can see
in lttng_session_metadata_statedump().

lttng_event_recorder_metadata_statedump_ctf_1_8() now receives an event
recorder instead of an event because that's all it needs to write TSDL.

The common lttng_session_metadata_statedump(),
lttng_channel_metadata_statedump(), and
lttng_event_recorder_metadata_statedump() deal with common steps
such as:

• Making sure the session is active.

• Calling lttng_metadata_begin() and lttng_metadata_end() so that any
  lttng_metadata_print*() occurs between those.

• Reading and setting the `metadata_dumped` fields.

• Ensuring the channel and event recorder types.

Change-Id: I6b1170df3218087df4e8e0763a950216de46ff46
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>