Test: Add an enum that uses multiple ranges for an entry
In order to exercise more of the metadata generation code, add an
enumeration that defines an entry associtated to multiple ranges.

This new enum mimicks `proto_transport_header_type` used by the
network instrumentation which uncovered bugs in the CTF2 generation
code when assembling the set of ranges associated to a given
enumeration entry.

A test in the lttng-tools project
(tests/regression/tools/trace-format/test_trace_format_ctf2_kernel.py)
uses this test module to validate that the resulting trace is
legal and correct.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I3066d1ba79922279e18c325eee84415baf5d94ae