Fix: src/init-enum-desc-sorted-entries.c: set first `is_auto` to 0
When the very first entry has the `is_auto` option, its real beginning
and end values are both 0.

Without this patch, they're both 1.

Keep the _next beginning_ value instead of the last end value so as to
set the value of the first entry correctly. For a given entry, the next
beginning value is always its end value plus one.

Change-Id: I246d3f635a70d80c65385eb6edafcc8aa412a0d2
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>