Fix: trace_balance_dirty_pages in Linux v6.14.2
commit 12169d8a25c5 ("Fix: trace_balance_dirty_pages parameters changed in Linux v6.14.2")
introduces a use of cgroup_ino(), which appears to be unavailable on
kernel configurations with CONFIG_CGROUP_WRITEBACK=n.

Remove this field from the event prior to kernel v6.15.0 so we don't
have to introduce an equivalent to __trace_wb_assign_cgroup to
work-around the fact that cgroup_ino is not available in older kernels
when CONFIG_CGROUP_WRITEBACK=n.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ie123c4446199305ee0b591bc7b94e711cb956fb1