Fix: metadata-ctf2: enum mappings are not serialized
On RHEL8 running kernel ""4.18.0-553.27.1.el8_10.x86_64", the
LTTng-tools "trace-format" test fails when comparing babeltrace's
"pretty" and "details" outputs between traces recorded in CTF 1.8 or
CTF2 format.

The output (not included for brevity) points to the declaration of
enumerations not matching in the output of the `sink.text.details`
output.

Inspecting the metadata of the CTF2 trace reveals that the "mappings"
dictionary of the various enum field classes are all empty, e.g.

  {
    "name": "enum300",
    "field-class": {
      "type": "fixed-length-signed-integer",
      "length": 32,
      "byte-order": "little-endian",
      "alignment": 8,
      "mappings": {}
    }
  },

Conversely, the same enumeration in CTF 1.8 format has the following
declaration:

  enum : integer { size = 32; align = 8; signed = 1; encoding = none; base = 10; } {
    "multiple ranges" = 0 ... 99,
    "entry a" = 100,
    "multiple ranges" = 101 ... 199,
    "entry b" = 200,
    "multiple ranges" = 201 ... 299,
    "entry c" = 300,
    "multiple ranges" = 301 ... 1023,
  } _enum300;

Looking at write_int_fc_mappings(), we see that the index used to
iterate over the field class' mappings is not initialized.

Assigning it to zero fixes the problem and allows the test to succeed on
that platform.

The introduction of CONFIG_INIT_STACK_ALL_ZERO in 5.9+ probably explains
why the problem is only visible on this older kernel.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ibb69c81091ccad6328a3e45fc01348b87dc2b7d4