Fix: Sample discarded packet counters per sub-buffer
Observed issue
==============

In LTTng-UST, which mirrors the implementation of the LTTng-modules ring
buffer in fully preemptible context, we observe the following:

In LTTng-tools,
`./tests/regression/ust/high-throughput/test_high_throughput` fails
with an unlikely number of discarded events.

E.g.

    Expected 20000000. Dropped 442682350. Recorded 5846419
    /tmp/tmp.test_high_throughput_tmp_file1.rC8QCO. Total 448528769

Cause
=====

The sub-buffer finalization order is not guaranteed to be sequential.
If a producer (traced application) is interrupted between the reserve
and commit, a sub-buffer could fill with other events but remain
open. In the meantime, further events could fill subsequent
sub-buffers that could be closed.

E.g.

1. Packet 2 closed (records lost sampled at switch = 0)
2. Packet 3 closed (records lost sampled at switch = 0)
3. Packet 4 closed (records lost sampled at switch = 0)
4. Packet 1 closed (records lost sampled at switch = 459)
5. Packet 2 closed (records lost sampled at switch = 459)

If, in the above example, events started to only be discarded after
packet 4 is closed due to wrap-around, sub-buffer 1 would close with a
higher events discarded value than in sub-buffers 2, 3, and 4.

The `events_discarded` value of a sub-buffer should be the number of
packets discarded in between the start timestamp and end timestamp of
the packet. Trace readers expect that the value of `events_discarded`
increases from sub-buffer to sub-buffer. If the events_discarded value
of sub-buffer N is ever less than that of sub-buffer N-1, it is
interpreted as a wrap-around in the counter.

This inconsistency causes Babeltrace to report many more discarded
events that are actually occurring.

A similar scenario may happen within LTTng-modules if an interrupt nests
between a reserve and commit, thus causing the events_discarded counters
to go backward between two consecutive packets within a stream.

Solution
========

Sample the sum of the records lost into the `end_events_discarded` on a
per sub-buffer basis, using padding within the per sub-buffer cc_cold
structure.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I1710d7d9e81544698f19a3de70e04323ad6ae4f3